requirements.txt
py-cord
sqlalchemy
psycopg2-binary
google-generativeai
python-dotenv
